move dialog to the "dialogs" folder
