+++ src/app/app.module.ts
@@ -64,7 +64,7 @@ import { NodeRestoreDirective } from './common/directives/node-restore.directive
 import { NodePermanentDeleteDirective } from './common/directives/node-permanent-delete.directive';
 import { NodeUnshareDirective } from './common/directives/node-unshare.directive';
 import { NodeVersionsDirective } from './common/directives/node-versions.directive';
-import { VersionManagerDialogAdapterComponent } from './components/versions-dialog/version-manager-dialog-adapter.component';
+import { VersionManagerDialogAdapterComponent } from './dialogs/versions-dialog/version-manager-dialog-adapter.component';
 import { BrowsingFilesService } from './common/services/browsing-files.service';
 import { ContentManagementService } from './common/services/content-management.service';
 import { NodeActionsService } from './common/services/node-actions.service';
