+++ src/app/dialogs/versions-dialog/version-manager-dialog-adapter.component.html
