+++ src/app/store/actions.ts
@@ -4,3 +4,4 @@ export * from './actions/logo-path.action';
 export * from './actions/node.action';
 export * from './actions/snackbar.action';
 export * from './actions/router.action';
+export * from './actions/viewer.action';
